pyproject.toml
[project]
name = "cnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "inference-sdk>=0.50.3",
    "logging>=0.4.9.6",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pyright>=1.1.400",
    "python-dotenv>=1.1.0",
    "roboflow>=1.1.63",
    "ruff>=0.11.9",
    "tools>=1.0.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "ultralytics>=8.3.131",
]
